Sadalīju python skriptus uz simpleTest.py un OpticalFlow.py
Tagad  simpleTest paliek tikai lai pārbaudīt pieslēgumu pie coppelia sima
